Move structure and add directory_path arg
